+++ cmd/psweb/handlers.go
@@ -354,7 +354,7 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 	ln.ListUnspent(cl, &utxosBTC, 1)
 
 	var utxosLBTC []liquid.UTXO
-	liquid.ListUnspent(&utxosLBTC)
+	liquid.ListUnspent(&utxosLBTC, elementsBitcoinId)
 
 	// to find a channel for swap-out
 	maxLocalBalance := uint64(0)
@@ -435,7 +435,7 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 	swapFeeReserveBTC := uint64(math.Ceil(bitcoinFeeRate * 350))
 
 	// arbitrary haircut to avoid 'no matching outgoing channel available'
-	maxLiquidSwapIn := min(int64(satAmount)-int64(swapFeeReserveLBTC()), int64(maxRemoteBalance)-10000)
+	maxLiquidSwapIn := min(int64(satAmount)-int64(swapFeeReserveLBTC(len(utxosLBTC))), int64(maxRemoteBalance)-10000)
 	if maxLiquidSwapIn < 100_000 {
 		maxLiquidSwapIn = 0
 	}
@@ -451,7 +451,7 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 		} else {
 			peerLiquidBalance = formatWithThousandSeparators(ptr.Amount)
 		}
-		maxLiquidSwapOut = uint64(max(0, min(int64(maxLocalBalance)-SWAP_OUT_CHANNEL_RESERVE, int64(ptr.Amount)-int64(swapFeeReserveLBTC()))))
+		maxLiquidSwapOut = uint64(max(0, min(int64(maxLocalBalance)-SWAP_OUT_CHANNEL_RESERVE, int64(ptr.Amount)-int64(swapFeeReserveLBTC(1)))))
 	} else {
 		maxLiquidSwapOut = uint64(max(0, int64(maxLocalBalance)-SWAP_OUT_CHANNEL_RESERVE))
 	}
@@ -577,8 +577,6 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 		KeysendSats             uint64
 		OutputsBTC              *[]ln.UTXO
 		OutputsLBTC             *[]liquid.UTXO
-		ReserveLBTC             uint64
-		ReserveBTC              uint64
 		HasInboundFees          bool
 		PeerBitcoinBalance      string // "" means no data
 		MaxBitcoinSwapOut       uint64
@@ -633,8 +631,6 @@ func peerHandler(w http.ResponseWriter, r *http.Request) {
 		KeysendSats:             keysendSats,
 		OutputsBTC:              &utxosBTC,
 		OutputsLBTC:             &utxosLBTC,
-		ReserveLBTC:             swapFeeReserveLBTC(),
-		ReserveBTC:              swapFeeReserveBTC,
 		HasInboundFees:          ln.HasInboundFees(),
 		PeerBitcoinBalance:      peerBitcoinBalance,
 		MaxBitcoinSwapOut:       maxBitcoinSwapOut,
@@ -1775,7 +1771,7 @@ func liquidHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	walletInfo, err := liquid.GetWalletInfo(config.Config.ElementsWallet)
+	walletInfo, err := liquid.GetWalletInfo()
 	if err != nil {
 		redirectWithError(w, r, "/?", err)
 		return
@@ -2424,7 +2420,7 @@ func submitHandler(w http.ResponseWriter, r *http.Request) {
 				addressType = "bech32m"
 			}
 
-			addr, err := liquid.GetNewAddress(label, addressType, config.Config.ElementsWallet)
+			addr, err := liquid.GetNewAddress(label, addressType)
 			if err != nil {
 				redirectWithError(w, r, "/liquid?", err)
 				return
@@ -2767,9 +2763,8 @@ func loadingHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func backupHandler(w http.ResponseWriter, r *http.Request) {
-	wallet := config.Config.ElementsWallet
 	// returns .bak with the name of the wallet
-	if fileName, err := liquid.BackupAndZip(wallet); err == nil {
+	if fileName, err := liquid.BackupAndZip(); err == nil {
 		// Set the Content-Disposition header to suggest a filename
 		w.Header().Set("Content-Disposition", fmt.Sprintf("attachment; filename=%s", fileName))
 		// Serve the file for download